D2D_guide_backend/requirement.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
sqlparse==0.5.1
psycopg==3.2.1
django-cors-headers==4.4.0
django-filter==24.3
